backend/requirements-gcp.txt
# Google Cloud App Engine Requirements
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Werkzeug==2.3.7
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
python-socketio==5.9.0
python-engineio==4.7.1
bleach==6.1.0
flask-cors==4.0.0
cryptography>=41.0.0

# Production dependencies
gunicorn==21.2.0

# Additional dependencies
Pillow>=10.0.0
requests>=2.31.0
bleach>=6.0.0

# Google Cloud dependencies
google-cloud-bigquery>=3.11.0
google-api-core>=2.11.0
google-cloud-sql-connector>=1.0.0
google-cloud-secret-manager>=2.16.0

# PostgreSQL for Cloud SQL
psycopg2-binary>=2.9.9

# Cloud Storage (optional)
google-cloud-storage>=2.10.0